Price Fix v.2

+++ db/migrate/20130729051510_add_price_to_listings.rb
@@ -1,5 +1,5 @@
 class AddPriceToListings < ActiveRecord::Migration
   def change
-    add_column :listings, :price, :integer
+    add_column :listings, :price, :decimal
   end
 end
